requirements.txt
Django==5.1.4
djangorestframework>=3.10,<4.0
python-decouple>=3.0,<4.0

psycopg2>=2.0,<3.0

